requirements.deploy.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
decorator==5.1.1
dm-tree==0.1.8
flatbuffers==24.3.25
gast==0.4.0
gin-config==0.5.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
gym==0.23.0
gym-notices==0.0.8
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
keras==2.11.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.0
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.0
pydantic_core==0.40.0
pygame==2.1.3
Pygments==2.18.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.19.0
termcolor==2.5.0
tf-agents==0.16.0
tf-keras==2.15.0
typing_extensions==4.5.0
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
